third_party/blink/web_tests/http/tests/credentialmanager/publickeycredential-basics.html

<!DOCTYPE html>
<title>Credential Manager: PublicKeyCredential basics.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../js-test-resources/document-interface-broker-helpers.js"></script>
<script src="/gen/layout_test_data/mojo/public/js/mojo_bindings_lite.js"></script>
<script src="/gen/mojo/public/mojom/base/unguessable_token.mojom-lite.js"></script>
<script src="/gen/mojo/public/mojom/base/big_buffer.mojom-lite.js"></script>
<script src="/gen/mojo/public/mojom/base/string16.mojom-lite.js"></script>
<script src="/gen/mojo/public/mojom/base/time.mojom-lite.js"></script>
<script src="/gen/url/mojom/origin.mojom-lite.js"></script>
<script src="/gen/url/mojom/url.mojom-lite.js"></script>
<script src="/gen/third_party/blink/public/mojom/credentialmanager/credential_manager.mojom-lite.js"></script>
<script src="/gen/third_party/blink/public/mojom/webauthn/authenticator.mojom-lite.js"></script>
<script src="/gen/third_party/blink/public/mojom/webauthn/virtual_authenticator.mojom-lite.js"></script>
<script src="resources/virtual-navigator-credentials.js"></script>
<script>

if (document.location.host != "subdomain.example.test:8443") {
  document.location = "https://subdomain.example.test:8443/credentialmanager/publickeycredential-basics.html";
  promise_test(_ => new Promise(_ => {}), "Stall tests on the wrong host.");
}

promise_test(async _ => {
  assert_false(await PublicKeyCredential.isUserVerifyingPlatformAuthenticatorAvailable());
}, "isUserVerifyingPlatformAuthenticatorAvailable() is false if there are no platform authenticators.");

promise_test(async _ => {
  await navigator.credentials.test.createAuthenticator({
    attachment: blink.mojom.AuthenticatorAttachment.PLATFORM
  });
  assert_equals((await navigator.credentials.test.authenticators()).length, 1);
  assert_true(await PublicKeyCredential.isUserVerifyingPlatformAuthenticatorAvailable());
}, "isUserVerifyingPlatformAuthenticatorAvailable() is true if there are platform authenticators.");

promise_test(t => {
  return navigator.credentials.test.clearAuthenticators();
}, "Clean up testing environment.");
</script>
